move to data classes to model package
