+++ app/src/main/java/com/owncloud/android/ui/dialog/conflict/model/ConflictDialogType.kt
@@ -1,11 +1,11 @@
 /*
  * Nextcloud - Android Client
  *
- * SPDX-FileCopyrightText: 2024 Alper Ozturk <alper.ozturk@nextcloud.com>
+ * SPDX-FileCopyrightText: 2026 Alper Ozturk <alper.ozturk@nextcloud.com>
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-package com.owncloud.android.ui.dialog.parcel
+package com.owncloud.android.ui.dialog.conflict.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
@@ -23,14 +23,3 @@ sealed interface ConflictDialogType : Parcelable {
     @Parcelize
     data class Normal(override val dialogTitle: String, override val data: ConflictDialogData) : ConflictDialogType
 }
-
-@Parcelize
-data class ConflictDialogData(
-    val headline: String,
-    val description: String,
-    val localFile: ConflictFileData,
-    val serverFile: ConflictFileData
-) : Parcelable
-
-@Parcelize
-data class ConflictFileData(val title: String, val timestamp: String, val fileSize: String) : Parcelable
